+++ mkchromecast/__init__.py
@@ -48,6 +48,23 @@ formatter_class=RawTextHelpFormatter
 )
 
 parser.add_argument(
+'--alsa-device',
+type=str,
+default=None,
+help=
+'''
+Set the ALSA device name. This option is useful when you are using pure
+ALSA in your system.
+
+Example:
+    python mkchromecast.py --encoder-backend ffmpeg --alsa-device hw:2,1
+
+This option only works for the ffmpeg and avconv backends. It is not useful for
+pulseaudio users.
+'''
+)
+
+parser.add_argument(
 '-b',
 '--bitrate',
 type=int,
@@ -66,7 +83,6 @@ node:
 
 This option works with all backends. The example above sets the average
 bitrate to 128k.
-
 '''
 )
 
@@ -109,7 +125,6 @@ Possible codecs:
     - flac [HQ]     Free Lossless Audio Codec
 
 This option only works for the ffmpeg, avconv and parec backends.
-
 '''
 )
 
@@ -117,7 +132,7 @@ parser.add_argument(
 '--config',
 action='store_true',
 help='''
-Use this option to connect from configuration file
+Use this option to connect from configuration file.
 '''
 )
 
@@ -125,7 +140,7 @@ parser.add_argument(
 '--debug',
 action='store_true',
 help='''
-Option for debugging purposes
+Option for debugging purposes.
 '''
 )
 
@@ -134,7 +149,7 @@ parser.add_argument(
 '--discover',
 action='store_true',
 help='''
-Use this option if you want to know the friendly name of a Google Cast device
+Use this option if you want to know the friendly name of a Google Cast device.
 '''
 )
 
@@ -153,7 +168,6 @@ Possible backends:
 
 Example:
     python mkchromecast.py --encoder-backend ffmpeg
-
 '''
 )
 
@@ -162,7 +176,7 @@ parser.add_argument(
 '--name',
 action='store_true',
 help='''
-Use this option if you know the name of the Google Cast you want to connect
+Use this option if you know the name of the Google Cast you want to connect.
 '''
 )
 
@@ -180,7 +194,7 @@ parser.add_argument(
 action='store_true',
 help='''
 When the application fails, and you have no audio in your computer, use this
-option to reset the computer's audio
+option to reset the computer's audio.
 '''
 )
 
@@ -188,7 +202,7 @@ parser.add_argument(
 '--reboot',
 action='store_true',
 help='''
-Reboot the Google Cast device
+Reboot the Google Cast device.
 '''
 )
 
@@ -197,7 +211,7 @@ parser.add_argument(
 '--select-cc',
 action='store_true',
 help='''
-If you have more than one Google Cast device use this option
+If you have more than one Google Cast device use this option.
 '''
 )
 
@@ -236,7 +250,6 @@ Which sample rate to use?
     - 22050Hz: sampling rate of audio quality of AM radio.
 
 For more information see: http://wiki.audacityteam.org/wiki/Sample_Rates.
-
 '''
 )
 
@@ -245,7 +258,7 @@ parser.add_argument(
 '--tray',
 action='store_true',
 help='''
-This option let you launch mkchromecast as a systray menu (beta)
+This option let you launch mkchromecast as a systray menu (beta).
 '''
 )
 
@@ -262,7 +275,6 @@ This will execute for you:
 
     git pull --all
     git fetch -p
-
 """
 )
 
@@ -281,7 +293,8 @@ default=False,
 help='''
 This option lets you control the volume of your Google Cast Devices. Use the
 'u' and 'd' keys to perform volume up and volume down actions respectively. Note
-that to kill the application using this option, you need to press the 'q' key.
+that to kill the application using this option, you need to press the 'q' key
+or 'Ctrl-c'.
 '''
 )
 
@@ -298,7 +311,6 @@ Example:
 
 As I don't own a Google Cast for TVs, I cannot test this correctly. But in
 principle it should work.
-
 '''
 )
 
@@ -310,8 +322,10 @@ Guess the platform
 platform = platform.system()
 
 """
-Assigment of args to variables
+Assignment of args to variables
 """
+adevice = args.alsa_device
+print(adevice)
 tray = args.tray
 if tray == True:
     select_cc = True